backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
boto3==1.35.87
python-dotenv==0.19.0
werkzeug==2.3.7 